ovmm: Break dependency from hypervisors to core
