+++ openvmm/openvmm_core/src/hypervisor_backend.rs
@@ -89,42 +89,3 @@ impl ResolvedHypervisorBackend {
 impl CanResolveTo<ResolvedHypervisorBackend> for HypervisorKind {
     type Input<'a> = ();
 }
-
-/// Registers hypervisor backend probes for auto-detection.
-///
-/// Each entry is a unit struct implementing
-/// [`HypervisorProbe`](hypervisor_resources::HypervisorProbe).
-///
-/// Probes are checked in registration order when auto-detecting the
-/// hypervisor, so register them from highest to lowest priority.
-///
-/// Resource resolvers should be registered separately via
-/// [`vm_resource::register_static_resolvers!`].
-///
-/// # Example
-///
-/// ```ignore
-/// openvmm_core::register_hypervisor_probes! {
-///     #[cfg(all(target_os = "linux", feature = "virt_kvm", guest_is_native))]
-///     openvmm_hypervisors::kvm::KvmProbe,
-/// }
-/// ```
-#[macro_export]
-macro_rules! register_hypervisor_probes {
-    {} => {};
-    { $( $(#[$a:meta])* $probe:path ),+ $(,)? } => {
-        $(
-        $(#[$a])*
-        const _: () = {
-            static PROBE_INSTANCE: $probe = $probe;
-
-            #[hypervisor_resources::private::linkme::distributed_slice(
-                hypervisor_resources::private::HYPERVISOR_PROBES
-            )]
-            #[linkme(crate = hypervisor_resources::private::linkme)]
-            static PROBE: Option<&'static dyn hypervisor_resources::HypervisorProbe> =
-                Some(&PROBE_INSTANCE);
-        };
-        )*
-    };
-}
