+++ openvmm/openvmm_hypervisors/src/lib.rs
@@ -4,7 +4,7 @@
 //! Hypervisor backend implementations for OpenVMM.
 //!
 //! Each submodule provides a [`HypervisorProbe`](hypervisor_resources::HypervisorProbe)
-//! implementation and a resource resolver for the corresponding handle type.
+//! implementation for the corresponding hypervisor backend.
 //!
 //! Probes are registered here via `register_hypervisor_probes!`. Resource
 //! resolvers are registered separately in `openvmm_resources`.
@@ -17,7 +17,7 @@ pub mod mshv;
 pub mod whp;
 
 // Register probes for auto-detection (checked in this order).
-openvmm_core::register_hypervisor_probes! {
+hypervisor_resources::register_hypervisor_probes! {
     #[cfg(all(target_os = "linux", feature = "virt_mshv", guest_is_native))]
     mshv::MshvProbe,
