+++ openvmm/openvmm_hypervisors/src/whp.rs
@@ -7,7 +7,6 @@
 
 use hypervisor_resources::HypervisorKind;
 use hypervisor_resources::WhpHandle;
-use openvmm_core::hypervisor_backend::ResolvedHypervisorBackend;
 use vm_resource::Resource;
 
 /// WHP probe for auto-detection.
@@ -55,20 +54,3 @@ fn parse_bool_param(key: &str, val: &str) -> anyhow::Result<bool> {
         _ => anyhow::bail!("invalid boolean value for {key}: {val}"),
     }
 }
-
-/// WHP resource resolver.
-pub struct WhpResolver;
-
-impl vm_resource::ResolveResource<HypervisorKind, WhpHandle> for WhpResolver {
-    type Output = ResolvedHypervisorBackend;
-    type Error = std::convert::Infallible;
-
-    fn resolve(&self, resource: WhpHandle, _input: ()) -> Result<Self::Output, Self::Error> {
-        Ok(ResolvedHypervisorBackend::new(virt_whp::Whp {
-            user_mode_apic: resource.user_mode_apic,
-            offload_enlightenments: resource.offload_enlightenments,
-        }))
-    }
-}
-
-vm_resource::declare_static_resolver!(WhpResolver, (HypervisorKind, WhpHandle),);
